refactor: recoil 상태 폴더 구조 수정
