+++ src/states/cartProducts/selectors/index.ts
@@ -1,15 +1,6 @@
-import { atom, selector, selectorFamily } from 'recoil';
+import { selector, selectorFamily } from 'recoil';
 
-import { localStorageEffect } from './atomEffects';
-
-import { CART_STORAGE_ID } from '../constants/storage';
-import type { CartProduct } from '../types/product';
-
-export const cartProductState = atom<CartProduct[]>({
-  key: 'cartProductState',
-  default: [],
-  effects: [localStorageEffect(CART_STORAGE_ID)],
-});
+import { cartProductState } from '../atoms';
 
 export const cartProductCountState = selector({
   key: 'cartProductCountState',
